Add dependencies needed for standalone ddk_um build.
ddk_um requires specific static and Java libraries which
are not compiled even when included in PRODUCT_PACKAGES.
By using LOCAL_ADDITIONAL_DEPENDENCIES, it is possible to
maintain these dependencies in the OUT directory. If your
module needs to depend on anything else that isn't actually
built into it, you can add those make targets to LOCAL_ADDITIONAL_DEPENDENCIES.
This is usually a workaround for dependencies that aren't created automatically.

Signed-off-by: Andrii <andrii_chepurnyi@epam.com>
Acked-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>